CASMINST-1901: disable remote service repos
Suse has a notion of 'service' repos that are not treated like
regular repos.  We want these to be disabled by default so that
airgapped systems do not try to reach out to suse.com during repo
queries or other repo operations.

This commit adds a command to the cleanup-all-repos() function to
disable all remote service repos.  cleanup-all-repos() is called
for k8s and storage node builds, but not the ncn-common build.  This
is OK/desirable as the common image is an intermediate image and not
used directly.