Install older libfreebl3 rpms
Upstream suse repos were updated with newer versions of the
libfreebl3 rpms.  libfreebl3 is needed by the kernel.  Since
we are using an older version of the kernel, it needs an older
version of libfreebl3.

Add libfreebl3 to cms-rpms and pin it to the required version.
Add `--oldpackage` to the zypper install to allow packages downgrades.