MTL-1476 Fix Builds
SLE_HPC-release is handled in a provisioner for pit-common. The
node-image-build repo handles this RPM differently than
cray-pre-install-toolkit, and having this here actually breaks the
zypper command due to a conflict.

cray-pre-install-toolkit already has this package specified in its
config.xml file, it doesn't need to have it in csm-rpms.

Removing this will maintain the HPC integrity while enabling MTL-1476
builds.