Remove non-context files
All package files in here should align to the context division in
node-images, package files shouldn't group types of RPMs but rather
contexts.

This helps map the following:
- base is always installed in common
- google is always and only installed in google
- metal is always and only installed in metal

The cms and firmware package files didn't convey context, a newcomer
might not know whether they should be used in certain layers or not.