MTL-1865 Vagrant depends on `sudo` (#533)
The PIT image is failing to package in vagrant due to vagrant depending
n `sudo` when halting the machine, despite running as `root` vagrant
will always invoke `sudo /usr/bin/systemctl poweroff &`. Since `sudo` is
missing, that `poweroff` command fails.

This is not affecting the NCN common build because `sudo` is installed,
it's required by cloud-init and cray-sdu-rda.

This change adds the `sudo` package to both images, ensuring that we pin
it and track it instead of relying on `cloud-init` to require it.