MTL-1910 `pre-install-toolkit` is an NCN image
This change makes the pre-install-toolkit inherit from the
`non-compute-common` layer, removing the need to update two places for packages.
By not having to update two place we can avoid accidents, such as
updating the kernel for the NCN and forgetting to update the PIT. Other
packages like `craycli` also fail to be updated in all the necessary
places, sometimes.

- New packages to the PIT: `ceph-common`, `cfs*`, `cos`,
  `dracut-metal-mdsquash`, and `goss-servers` (not harmful)
- New packages to storage-ceph: `loftsman` and `manifestgen` now live in
  node-images-common
- Removed duplicates: `kernel-*`, `python*`, `craycli`, `hpe-csm-*`, (virtually)
  every OS package (e.g. `vim`, libraries, and more), `biosdevname`, `grub2*`,
  `mft`