MTL-2068 Pin `cri-tools`
`cri-tools-1.26.0-0` is now available, but it is incompatible with our
version of `containerd`. Since `cri-tools` was not pinned, the newer
`cri-tools` was being pulled in and causing NCN image builds to fail.

This pins the `cri-tools` package to the version used in our most
recent, working NCN image build (`1.25.0`). Without this pin, our
Kubernetes NCN image will fail to build.