Cleanup sles15-base; lock kernel in base layer
This removes the kernel packages and python310 from csm-rpms, these can in-fact be
installed with a version pin in the base layer while also being locked
by Zypper. In this way, we can remove all of the duplicated kernel
version cleanup and lock code that exists in application, compute, and
ncn-common.

***NOTE*** Python3.6 was left in csm-rpms because I wanted to leave the
management of the legacy Python version here, and only manage desired
Pythons from the base image.

This moves some of the packages installed in sles15-base into csm-rpms
so that they may be fully accounted for. sles15-base should only have
packages that are absolutley required for the base layer.