New repsotiroy for `cloud-init`
This adds a new repository that provides `cloud-init>21.4` for SLES.

The new hypervisor image can pull in native `cloud-init` and doesn't
need our fork to run, so it might as well use the latest/greatest.

Our other images using `cloud-init` will have to rely on `21.4` until
if/when our fork is no longer needed or is building `>21.4` (MTL-2092).